List the required dependencies in package.json
We're attempting to reduce our dependency footprint as much as possible.
Therefore, we're depending on lodash but only requiring 1 or 2 methods
from it for now. Hopefully we can achieve a build that will only pick
those.
We really need a uuid library. As for rxjs, it will be addressed later
on.