Change the way we consume _.omit
The TS compiler did not agree with the way we attempted to consume
omit as a single lodash method. The way it works is not compatible
with ES2015 modules: it's a non-forward-compatible CommonJS
implementation.
We'll have to find another way to package just this method, or to
reimplement it locally if we can't.