Add a query method to our ModelResource class
The query method allows us to perform GETs on collections and
stores the responses for each specific query. It also puts the
received resources into the store.
This is a somewhat complicated method as it has to keep a tight
grip over its slice of the state to only request the data it needs
and *never* *accidentally* rely on outdated data.